add: Added set activity, presence and music handler events
